Add pyproject and restructure tests
